+++ miracle/api/tests.py
@@ -2,10 +2,6 @@ import json
 
 from jwcrypto.jwk import JWK
 
-from miracle.api.views import (
-    check_user,
-)
-
 CORS_HEADERS = {
     'Access-Control-Allow-Origin',
     'Access-Control-Max-Age',
@@ -14,47 +10,28 @@ CORS_HEADERS = {
 }
 
 
-def test_check_user():
-    assert not check_user(None)
-    assert not check_user('')
-    assert not check_user(b'')
-    assert not check_user(b'ab')
-    assert not check_user(('a' * 40).encode('ascii'))
-    assert not check_user(b'abcd?')
-    assert not check_user(b'abcd\xfe')
-    assert check_user('abc')
-    assert check_user(b'abc')
-    assert check_user('fae006df902d4809aaddb176b6bdf8dd')
-    assert check_user(b'fae006df902d4809aaddb176b6bdf8dd')
-    assert check_user('fae006df-902d-4809-aadd-b176b6bdf8dd')
-    assert check_user(b'fae006df-902d-4809-aadd-b176b6bdf8dd')
-
-
 def test_jwk(app, stats):
-    res = app.get('/v1/jwk', status=200)
+    res = app.get('/v2/jwk', status=200)
     assert set(res.json.keys()) == {'n', 'e', 'kty'}
     assert res.json['kty'] == 'RSA'
     assert 'Strict-Transport-Security' in res.headers
     stats.check(timer=[
-        ('request', 1, ['path:v1.jwk', 'method:get', 'status:200']),
+        ('request', 1, ['path:v2.jwk', 'method:get', 'status:200']),
     ])
 
 
 def test_upload(app, crypto, stats):
     data = crypto.encrypt(json.dumps({'foo': 1}))
-    res = app.post('/v1/upload',
-                   data,
-                   headers={'Content-Type': 'text/plain',
-                            'X-User': b'abc'},
+    res = app.post('/v2/upload', data,
+                   headers={'Content-Type': 'text/plain'},
                    status=200)
     assert CORS_HEADERS - set(res.headers.keys()) == set()
     assert res.json == {'status': 'success'}
     assert 'Strict-Transport-Security' in res.headers
 
     data = crypto.encrypt(b'encrypted no json')
-    app.post('/v1/upload', data,
-             headers={'Content-Type': 'text/plain',
-                      'X-User': b'abc'},
+    app.post('/v2/upload', data,
+             headers={'Content-Type': 'text/plain'},
              status=200)
 
     stats.check(timer=[
@@ -63,24 +40,17 @@ def test_upload(app, crypto, stats):
 
 
 def test_upload_fail(app, stats):
-    app.post('/v1/upload', b'',
+    app.post('/v2/upload', b'',
              headers={'Content-Type': 'text/plain'},
              status=400)
-    app.post('/v1/upload', b'',
-             headers={'Content-Type': 'text/plain',
-                      'X-User': b'abc'},
+    app.post('/v2/upload', b'a' + b'0123456789' * 1024 * 1024,
+             headers={'Content-Type': 'text/plain'},
              status=400)
-    app.post('/v1/upload', b'a' + b'0123456789' * 1024 * 1024,
-             headers={'Content-Type': 'text/plain',
-                      'X-User': b'abc'},
+    app.post('/v2/upload', b'no json',
+             headers={'Content-Type': 'text/plain'},
              status=400)
-    app.post('/v1/upload', b'no json',
-             headers={'Content-Type': 'text/plain',
-                      'X-User': b'abc'},
-             status=400)
-    app.post('/v1/upload', 'no\xfejson'.encode('latin-1'),
-             headers={'Content-Type': 'text/plain',
-                      'X-User': b'abc'},
+    app.post('/v2/upload', 'no\xfejson'.encode('latin-1'),
+             headers={'Content-Type': 'text/plain'},
              status=400)
 
 
@@ -89,24 +59,22 @@ def test_upload_jwe(app, stats):
     jwk = JWK.generate(kty='RSA')
     data = app.app.registry.crypto.encrypt(
         b'{"key": "wrong"}', _public_jwk=jwk)
-    app.post('/v1/upload', data,
-             headers={'Content-Type': 'text/plain',
-                      'X-User': b'abc'},
+    app.post('/v2/upload', data,
+             headers={'Content-Type': 'text/plain'},
              status=200)
     # Encrypt the data with a wrong algorithm.
     data = app.app.registry.crypto.encrypt(
         b'{"alg": "wrong"}',
         _protected='{"alg":"RSA-OAEP","enc":"A128CBC-HS256"}')
-    app.post('/v1/upload', data,
-             headers={'Content-Type': 'text/plain',
-                      'X-User': b'abc'},
+    app.post('/v2/upload', data,
+             headers={'Content-Type': 'text/plain'},
              status=400)
 
 
 def test_head(app, stats):
     urls = [
-        '/v1/jwk',
-        '/v1/upload',
+        '/v2/jwk',
+        '/v2/upload',
     ]
     for url in urls:
         app.head(url, status=200)
@@ -115,8 +83,8 @@ def test_head(app, stats):
 def test_options(app, stats):
     urls = {
         # url: primary supported method
-        '/v1/jwk': 'GET',
-        '/v1/upload': 'POST',
+        '/v2/jwk': 'GET',
+        '/v2/upload': 'POST',
     }
     for url, method in urls.items():
         res = app.options(url, status=200)
@@ -127,8 +95,8 @@ def test_options(app, stats):
 def test_unsupported(app, stats):
     urls = {
         # url: unsupported method
-        '/v1/jwk': 'post',
-        '/v1/upload': 'get',
+        '/v2/jwk': 'post',
+        '/v2/upload': 'get',
     }
     for url, method in urls.items():
         app.delete(url, status=405)