+++ CHANGES.rst
@@ -5,6 +5,8 @@ Changelog
 2.0.0 (unreleased)
 ==================
 
+- #5: Update to v2 API, which moves the user id into the payload.
+
 - Remove delete and stats API endpoints.
 
 - Remove end date.